ocr-service/requirements.txt
numpy==1.26.4
opencv-python==4.6.0.66
pytesseract==0.3.10
flask
pillow